ouisync_plugin. framework added to the iOS workpace file It seems the framework that sghould be added automatically when building the iOS project, was not happening, and this was the reason (maybe?) for not finding the objkect defined in CallbackManager.
But after I manually added the framework file in the Runner on Xcode, it throwed an error on compilation saying two actions were trying to do the same:  adding the framework ouisync_plugin.
Once I removed the framework added manually, the compilations worked.

Also: changed the stryp style to non-global to prevent the symboils in the native files to be removed on compilation.