Refactored: PipeProvider.kt
The thread transfer between the UI thread, for invoking the method channel, and the thread for writing to the OutputStream of the pipe was simplified.
A extension function was added to the Context, to execute the invocation in the UI thread. This function is a higher order type, so no callback task is
used anymore; instead the getFileChunk function is passed as a parameter.