Bug fix: Intent type assignation removed; letting Android to determine it automatically.
Trying to guess the content type and using it for creating the intent was causing the preview function to not work as expected.
The assignation was removed, so Android takes care of it.

Also: in PipeProvider, the CONTENT_URI url was modified to not include / at the end; this way the file path received from the Flutter can be appenened
without any extra formatting in the main activity when creating the uri for the intents.